chore: remove outcome rows hook
