+++ app/components/UI/Predict/components/PredictGameDetailsContent/usePredictGameGroupOutcomes.test.ts
@@ -4,13 +4,17 @@ import type {
   PredictOutcomeGroup,
   PredictOutcomeToken,
 } from '../../types';
-import { usePredictGameGroupOutcomes } from './usePredictGameGroupOutcomes';
+import {
+  buildOutcomeCardModels,
+  usePredictGameGroupOutcomes,
+} from './usePredictGameGroupOutcomes';
 
 jest.mock('../../../../../../locales/i18n', () => ({
   strings: jest.fn((key: string) => {
     const translations: Record<string, string> = {
       'predict.sports_market_types.moneyline': 'Moneyline',
       'predict.sports_market_types.total_corners': 'Corners',
+      'predict.sports_market_types.soccer_player_goals': 'Goals',
     };
 
     return translations[key] ?? key;
@@ -79,6 +83,10 @@ const createGroup = (
 });
 
 describe('usePredictGameGroupOutcomes', () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
+
   it('returns empty values when no group is provided', () => {
     const { result } = renderHook(() => usePredictGameGroupOutcomes({}));
 
@@ -292,5 +300,255 @@ describe('usePredictGameGroupOutcomes', () => {
     expect(result.current.closedOutcomes).toHaveLength(4);
     expect(result.current.showResolvedSection).toBe(true);
     expect(result.current.hasPartialResolution).toBe(false);
+  });
+
+  it('builds open card models and de-duplicated token ids for subgrouped outcomes', () => {
+    const group = createGroup('game_lines', [
+      createCard({
+        key: 'moneyline',
+        outcomes: [
+          createOutcome({
+            id: 'moneyline-home',
+            sportsMarketType: 'moneyline',
+            tokens: [createToken({ id: 'ml-home', price: 0.6 })],
+          }),
+          createOutcome({
+            id: 'moneyline-away',
+            sportsMarketType: 'moneyline',
+            tokens: [createToken({ id: 'ml-away', price: 0.4 })],
+          }),
+        ],
+      }),
+      createCard({
+        key: 'soccer_player_goals-0',
+        title: 'Player One',
+        outcomes: [
+          createOutcome({
+            id: 'player-goals',
+            sportsMarketType: 'soccer_player_goals',
+            groupItemTitle: 'Player One: 1+ goals',
+            tokens: [
+              createToken({ id: 'player-over', shortTitle: 'Over' }),
+              createToken({ id: 'player-under', shortTitle: 'Under' }),
+            ],
+          }),
+        ],
+      }),
+      createCard({
+        key: 'total_corners',
+        outcomes: [
+          createOutcome({
+            id: 'corners-85',
+            sportsMarketType: 'total_corners',
+            groupItemTitle: 'Total Corners: O/U 8.5',
+            line: 8.5,
+            volume: 1000,
+            tokens: [
+              createToken({ id: 'corners-over-85' }),
+              createToken({ id: 'corners-under-85' }),
+            ],
+          }),
+          createOutcome({
+            id: 'corners-95',
+            sportsMarketType: 'total_corners',
+            groupItemTitle: 'Total Corners: O/U 9.5',
+            line: 9.5,
+            volume: 5000,
+            tokens: [
+              createToken({ id: 'corners-over-95' }),
+              createToken({ id: 'corners-under-95' }),
+            ],
+          }),
+        ],
+      }),
+      createCard({
+        key: 'duplicate-token-card',
+        title: 'Duplicate Token Card',
+        outcomes: [
+          createOutcome({
+            id: 'duplicate-token-outcome',
+            sportsMarketType: 'soccer_player_goals',
+            groupItemTitle: 'Duplicate Token Outcome',
+            tokens: [
+              createToken({ id: 'player-over', shortTitle: 'Over Again' }),
+              createToken({ id: 'player-under', shortTitle: 'Under Again' }),
+            ],
+          }),
+        ],
+      }),
+    ]);
+
+    const { result } = renderHook(() => usePredictGameGroupOutcomes({ group }));
+
+    expect(result.current.openCardModels).toEqual([
+      expect.objectContaining({
+        kind: 'moneyline',
+        key: 'moneyline',
+        title: 'Moneyline',
+        testID: 'game_lines-moneyline',
+      }),
+      expect.objectContaining({
+        kind: 'simple',
+        key: 'player-goals',
+        title: 'Player One',
+        testID: 'game_lines-soccer_player_goals-0',
+      }),
+      expect.objectContaining({
+        kind: 'line',
+        key: 'total_corners',
+        title: 'Corners',
+        testID: 'game_lines-total_corners',
+      }),
+      expect.objectContaining({
+        kind: 'simple',
+        key: 'duplicate-token-outcome',
+        title: 'Duplicate Token Card',
+        testID: 'game_lines-duplicate-token-card',
+      }),
+    ]);
+
+    expect(result.current.activeGroupTokenIds).toEqual([
+      'ml-home',
+      'ml-away',
+      'player-over',
+      'player-under',
+      'corners-over-85',
+      'corners-under-85',
+      'corners-over-95',
+      'corners-under-95',
+    ]);
+  });
+
+  it('keeps memoized references stable when the same group instance is rerendered', () => {
+    const group = createGroup('game_lines', [
+      createCard({
+        key: 'total_corners',
+        outcomes: [
+          createOutcome({
+            id: 'corners-95',
+            sportsMarketType: 'total_corners',
+            groupItemTitle: 'Total Corners: O/U 9.5',
+            line: 9.5,
+            tokens: [
+              createToken({ id: 'corners-over-95' }),
+              createToken({ id: 'corners-under-95' }),
+            ],
+          }),
+        ],
+      }),
+    ]);
+
+    const { result, rerender } = renderHook(
+      ({ currentGroup }) =>
+        usePredictGameGroupOutcomes({ group: currentGroup }),
+      { initialProps: { currentGroup: group } },
+    );
+
+    const firstCardModels = result.current.openCardModels;
+    const firstTokenIds = result.current.activeGroupTokenIds;
+
+    rerender({ currentGroup: group });
+
+    expect(result.current.openCardModels).toBe(firstCardModels);
+    expect(result.current.activeGroupTokenIds).toBe(firstTokenIds);
+  });
+
+  it('recomputes derived values when the group changes', () => {
+    const firstGroup = createGroup('game_lines', [
+      createCard({
+        key: 'moneyline',
+        outcomes: [
+          createOutcome({
+            id: 'moneyline-home',
+            sportsMarketType: 'moneyline',
+            tokens: [createToken({ id: 'ml-home', price: 0.6 })],
+          }),
+          createOutcome({
+            id: 'moneyline-away',
+            sportsMarketType: 'moneyline',
+            tokens: [createToken({ id: 'ml-away', price: 0.4 })],
+          }),
+        ],
+      }),
+    ]);
+    const secondGroup = createGroup('corners', [
+      createCard({
+        key: 'total_corners',
+        outcomes: [
+          createOutcome({
+            id: 'corners-85',
+            sportsMarketType: 'total_corners',
+            groupItemTitle: 'Total Corners: O/U 8.5',
+            line: 8.5,
+            tokens: [
+              createToken({ id: 'corners-over-85' }),
+              createToken({ id: 'corners-under-85' }),
+            ],
+          }),
+          createOutcome({
+            id: 'corners-95',
+            sportsMarketType: 'total_corners',
+            groupItemTitle: 'Total Corners: O/U 9.5',
+            line: 9.5,
+            tokens: [
+              createToken({ id: 'corners-over-95' }),
+              createToken({ id: 'corners-under-95' }),
+            ],
+          }),
+        ],
+      }),
+    ]);
+
+    const { result, rerender } = renderHook(
+      ({ currentGroup }) =>
+        usePredictGameGroupOutcomes({ group: currentGroup }),
+      { initialProps: { currentGroup: firstGroup } },
+    );
+
+    rerender({ currentGroup: secondGroup });
+
+    expect(result.current.openCardModels).toEqual([
+      expect.objectContaining({
+        kind: 'line',
+        key: 'total_corners',
+        title: 'Corners',
+        testID: 'corners-total_corners',
+      }),
+    ]);
+    expect(result.current.activeGroupTokenIds).toEqual([
+      'corners-over-85',
+      'corners-under-85',
+      'corners-over-95',
+      'corners-under-95',
+    ]);
+  });
+});
+
+describe('buildOutcomeCardModels', () => {
+  it('builds card models directly from a group', () => {
+    const group = createGroup('game_lines', [
+      createCard({
+        key: 'moneyline',
+        outcomes: [
+          createOutcome({
+            id: 'moneyline-home',
+            sportsMarketType: 'moneyline',
+            tokens: [createToken({ id: 'ml-home', price: 0.6 })],
+          }),
+          createOutcome({
+            id: 'moneyline-away',
+            sportsMarketType: 'moneyline',
+            tokens: [createToken({ id: 'ml-away', price: 0.4 })],
+          }),
+        ],
+      }),
+    ]);
+
+    expect(buildOutcomeCardModels(group)).toEqual([
+      expect.objectContaining({
+        kind: 'moneyline',
+        key: 'moneyline',
+      }),
+    ]);
   });
 });
